improved infrastructure, password security, mail drafting, added system metrics

+++ telefynaConfiguration/sendMail.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
 cd /var/www/html/telefynaConfiguration/;
-cat mailer.txt exports/$(date '+%Y-%m-%d')*.json exports/$(date '+%Y-%m-%d')*.txt  > content.txt;
-cat content.txt | msmtp --debug -a gmail apps@avventohome.org;
-rm exports/$(date '+%Y-%m-%d')*.json;
-rm exports/$(date '+%Y-%m-%d')*.txt;
-rm content.txt;
+FILES=exports/$(date '+%Y-%m-%d')*.txt
+if compgen -G "$FILES" > /dev/null; then
+    cat mailHeader.txt exports/$(date '+%Y-%m-%d')*.txt mailFooter.txt > content.txt;
+    cat content.txt | msmtp --debug -a gmail apps@avventohome.org;
+    rm content.txt;
+fi
 cd -;
